[canonicalise-hashes] Canonicalise hashes before comparing them
The replace and withdraw operations involve server-side hash
comparisons, so as to establish agreement between the client and the
server about the files that will be affected.  On the server side, the
hashes are generated with uppercase characters (see the Hashing
trait), but the client's hash is used unchanged, which means that the
client must also send hashes with uppercase characters for these
operations to be successful.  The relevant specification (RFC 8181)
does not mandate uppercase or lowercase characters in the hash, so
the hashes are now canonicalised before they are compared.  The
current behaviour causes problems for the rpki.net publication client,
which in at least some instances will send hashes that contain
lowercase characters.