Do not freeze versions if neither `writeRsync` nor `writeRrdp` is set.
The change of behaviour here is that if the publication server is not
configured to write either rrdp or rsync repository it will not freeze
versions in the database. That allows to have an instance that writes
publication messages in the DB but doesn't create repository files.
It can be useful for deployments without downtime.