Update repositories on disk on a fixed interval
Replace on-demand writing of repository files with a fixed schedule. The
previous behavior caused publication messages to wait for FS updates, because
the `PublicationService` was used as lock target.

Rather than fixing the locking code, update the FS on a fixed interval. Checking
for changes in the database is quite cheap, and having a fixed schedule leads to
simpler code.