Ensure all published objects are part of a delta when freezing version
Previously it was possible that an in-progress transaction would be
missing from a delta due to the non-transactional nature of sequence
number generation.

The main problem is that `object_log.id` may not be strictly
sequential, but we depend on this when querying a specific
version (since we only store the highest `object_log.id` per version).

Scenario:

- Client 1 starts publishing a million objects (getting ids 1
  ... 499_999, 500_001 ... 1_000_001)

- Client 2 publishes one object (getting id 500_000) and commits.

- Freeze version is called, creating a version with highest id
  500_000. The delta only contains the object from client 2, since
  client 1 has not committed yet.

- Client 1 commits, making objects 1 ... 499_999, 500_001
  ... 1_000_001 visible.

- Freeze version is called, creating a version with highest id
  1_000_001. The delta contains objects 500_001 to 1_000_001, missing
  the first 499_999 objects.

To solve this we now explicitly link `object_log` entries to the
version record ensuring even entries with lower `id` than
`last_log_entry_id` are part of the new delta.