Always update the database after generating files
In theory the generated file should have the same hash and size as the
previously generated file. However, if we ever change the formatting
of the generated files this is no longer the case. So always update
the information in the database with the hash/size of the generated
file.