Test that empty minimized delta does not lead to a new serial
After combining all operatings for a new delta the delta should not be
generated if the combined operations are empty. For example, if a
client publishes and withdraws an object in two separate requests no
new RRDP delta should be generated.

The same is true when an object is replaced and then the old object is
restored again.