Always randomize file names when generating snapshots and deltas
Snapshot and delta files are supposed to be immutable and can be
cached "forever". However, when the publication server starts up it
re-generates the needed snapshot and delta files. But there is no
guarantee that the files will have exactly the same content (for
example, due to code changes or layout changes).

So now we always randomize the filenames and generate a new
notification.xml file.

This breaks multi-instance mode where multiple instances are running
from a single database, since the notification.xml file will not
match on every instance which may cause trouble when fetching the
referenced files from another instance.

The solution is to use deterministic filenames based on the contents,
such as using the SHA-256 of the file in the filename. However, since
we do not want to have predictable file names (due to caching of 404
responses) it should also include a secret and use an algorithm such
as HmacSHA256 to generate the unique filename.

This will be part of a later commit.