Base RRDP file names on content hash and version specific secret
Instead of fully randomizing the file names they are now based on an
HMAC of the generated content. The secret is generated randomly for
every RRDP version and stored in the database.

Whenever the file is (re)created the filename is based on the HMAC
SHA-256. On re-creation the files should have the same name, unless
the contents are now different due to code changes in the publication
server.